+++ src/wavbreaker.h
@@ -45,7 +45,7 @@ struct MoodbarData_ {
 typedef struct TrackBreak_ TrackBreak;
 struct TrackBreak_ {
 	gboolean  write;
-	guint     offset;
+	gulong    offset;
 	gchar     *filename;
 	gchar     time[128];
 	gchar     duration[128];
@@ -74,7 +74,7 @@ void track_break_rename();
 void wavbreaker_write_files(char *dirname);
 GtkWidget *wavbreaker_get_main_window();
 gboolean open_file_arg( gpointer data);
-void track_break_add_offset( char* filename, guint offset);
+void track_break_add_offset(char *filename, gulong offset);
 void track_break_clear_list();
 void wavbreaker_quit();
